+++ simple_history/tests/models.py
@@ -4,10 +4,12 @@ import uuid
 from django.apps import apps
 from django.conf import settings
 from django.db import models
+from django.db.models.deletion import CASCADE
+from django.db.models.fields.related import ForeignKey
 from django.urls import reverse
 
 from simple_history import register
-from simple_history.models import HistoricalRecords
+from simple_history.models import HistoricalRecords, HistoricForeignKey
 
 from .custom_user.models import CustomUser as User
 from .external.models import AbstractExternal, AbstractExternal2, AbstractExternal3
@@ -23,6 +25,14 @@ class Poll(models.Model):
 
     def get_absolute_url(self):
         return reverse("poll-detail", kwargs={"pk": self.pk})
+
+
+class PollWithNonEditableField(models.Model):
+    question = models.CharField(max_length=200)
+    pub_date = models.DateTimeField("date published")
+    modified = models.DateTimeField(auto_now=True, editable=False)
+
+    history = HistoricalRecords()
 
 
 class PollWithUniqueQuestion(models.Model):
@@ -69,7 +79,7 @@ class PollWithExcludedFKField(models.Model):
 
 class AlternativePollManager(models.Manager):
     def get_queryset(self):
-        return super(AlternativePollManager, self).get_queryset().exclude(id=1)
+        return super().get_queryset().exclude(id=1)
 
 
 class PollWithAlternativeManager(models.Model):
@@ -102,13 +112,13 @@ class PollWithHistoricalIPAddress(models.Model):
 class CustomAttrNameForeignKey(models.ForeignKey):
     def __init__(self, *args, **kwargs):
         self.attr_name = kwargs.pop("attr_name", None)
-        super(CustomAttrNameForeignKey, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
 
     def get_attname(self):
-        return self.attr_name or super(CustomAttrNameForeignKey, self).get_attname()
+        return self.attr_name or super().get_attname()
 
     def deconstruct(self):
-        name, path, args, kwargs = super(CustomAttrNameForeignKey, self).deconstruct()
+        name, path, args, kwargs = super().deconstruct()
         if self.attr_name:
             kwargs["attr_name"] = self.attr_name
         return name, path, args, kwargs
@@ -122,15 +132,13 @@ class ModelWithCustomAttrForeignKey(models.Model):
 class CustomAttrNameOneToOneField(models.OneToOneField):
     def __init__(self, *args, **kwargs):
         self.attr_name = kwargs.pop("attr_name", None)
-        super(CustomAttrNameOneToOneField, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
 
     def get_attname(self):
-        return self.attr_name or super(CustomAttrNameOneToOneField, self).get_attname()
+        return self.attr_name or super().get_attname()
 
     def deconstruct(self):
-        name, path, args, kwargs = super(
-            CustomAttrNameOneToOneField, self
-        ).deconstruct()
+        name, path, args, kwargs = super().deconstruct()
         if self.attr_name:
             kwargs["attr_name"] = self.attr_name
         return name, path, args, kwargs
@@ -138,7 +146,7 @@ class CustomAttrNameOneToOneField(models.OneToOneField):
 
 class ModelWithCustomAttrOneToOneField(models.Model):
     poll = CustomAttrNameOneToOneField(Poll, models.CASCADE, attr_name="custom_poll")
-    history = HistoricalRecords(excluded_field_kwargs={"poll": set(["attr_name"])})
+    history = HistoricalRecords(excluded_field_kwargs={"poll": {"attr_name"}})
 
 
 class Temperature(models.Model):
@@ -189,15 +197,13 @@ class Voter(models.Model):
 class HistoricalRecordsVerbose(HistoricalRecords):
     def get_extra_fields(self, model, fields):
         def verbose_str(self):
-            return "%s changed by %s as of %s" % (
+            return "{} changed by {} as of {}".format(
                 self.history_object,
                 self.history_user,
                 self.history_date,
             )
 
-        extra_fields = super(HistoricalRecordsVerbose, self).get_extra_fields(
-            model, fields
-        )
+        extra_fields = super().get_extra_fields(model, fields)
         extra_fields["__str__"] = verbose_str
         return extra_fields
 
@@ -224,7 +230,7 @@ class Person(models.Model):
         if hasattr(self, "skip_history_when_saving"):
             raise RuntimeError("error while saving")
         else:
-            super(Person, self).save(*args, **kwargs)
+            super().save(*args, **kwargs)
 
 
 class FileModel(models.Model):
@@ -291,7 +297,9 @@ class State(models.Model):
 
 class Book(models.Model):
     isbn = models.CharField(max_length=15, primary_key=True)
-    history = HistoricalRecords(verbose_name="dead trees")
+    history = HistoricalRecords(
+        verbose_name="dead trees", verbose_name_plural="dead trees plural"
+    )
 
 
 class HardbackBook(Book):
@@ -308,6 +316,7 @@ class Library(models.Model):
 
     class Meta:
         verbose_name = "quiet please"
+        verbose_name_plural = "quiet please plural"
 
 
 class BaseModel(models.Model):
@@ -375,6 +384,14 @@ class UnicodeVerboseName(models.Model):
 
     class Meta:
         verbose_name = "\u570b"
+
+
+class UnicodeVerboseNamePlural(models.Model):
+    name = models.CharField(max_length=100)
+    history = HistoricalRecords()
+
+    class Meta:
+        verbose_name_plural = "\u570b"
 
 
 class CustomFKError(models.Model):
@@ -690,13 +707,11 @@ class OverrideModelNameAsString(models.Model):
 
 class OverrideModelNameAsCallable(models.Model):
     name = models.CharField(max_length=15, unique=True)
-    history = HistoricalRecords(custom_model_name=lambda x: "Audit{}".format(x))
+    history = HistoricalRecords(custom_model_name=lambda x: f"Audit{x}")
 
 
 class AbstractModelCallable1(models.Model):
-    history = HistoricalRecords(
-        inherit=True, custom_model_name=lambda x: "Audit{}".format(x)
-    )
+    history = HistoricalRecords(inherit=True, custom_model_name=lambda x: f"Audit{x}")
 
     class Meta:
         abstract = True
@@ -757,3 +772,85 @@ class ModelWithExcludedManyToMany(models.Model):
     name = models.CharField(max_length=15, unique=True)
     other = models.ManyToManyField(ManyToManyModelOther)
     history = HistoricalRecords(excluded_fields=["other"])
+
+
+class ModelWithSingleNoDBIndexUnique(models.Model):
+    name = models.CharField(max_length=15, unique=True, db_index=True)
+    name_keeps_index = models.CharField(max_length=15, unique=True, db_index=True)
+    history = HistoricalRecords(no_db_index="name")
+
+
+class ModelWithMultipleNoDBIndex(models.Model):
+    name = models.CharField(max_length=15, db_index=True)
+    name_keeps_index = models.CharField(max_length=15, db_index=True)
+    fk = models.ForeignKey(
+        "Library", on_delete=models.CASCADE, null=True, related_name="+"
+    )
+    fk_keeps_index = models.ForeignKey(
+        "Library", on_delete=models.CASCADE, null=True, related_name="+"
+    )
+    history = HistoricalRecords(no_db_index=["name", "fk", "other"])
+
+
+class TestOrganization(models.Model):
+    name = models.CharField(max_length=15, unique=True)
+
+
+class TestOrganizationWithHistory(models.Model):
+    name = models.CharField(max_length=15, unique=True)
+    history = HistoricalRecords()
+
+
+class TestParticipantToHistoricOrganization(models.Model):
+    """
+    Non-historic table foreign key to historic table.
+
+    In this case it should simply behave like ForeignKey because
+    the origin model (this one) cannot be historic, so foreign key
+    lookups are always "current".
+    """
+
+    name = models.CharField(max_length=15, unique=True)
+    organization = HistoricForeignKey(
+        TestOrganizationWithHistory, on_delete=CASCADE, related_name="participants"
+    )
+
+
+class TestHistoricParticipantToOrganization(models.Model):
+    """
+    Historic table foreign key to non-historic table.
+
+    In this case it should simply behave like ForeignKey because
+    the origin model (this one) can be historic but the target model
+    is not, so foreign key lookups are always "current".
+    """
+
+    name = models.CharField(max_length=15, unique=True)
+    organization = HistoricForeignKey(
+        TestOrganization, on_delete=CASCADE, related_name="participants"
+    )
+    history = HistoricalRecords()
+
+
+class TestHistoricParticipanToHistoricOrganization(models.Model):
+    """
+    Historic table foreign key to historic table.
+
+    In this case as_of queries on the origin model (this one)
+    or on the target model (the other one) will traverse the
+    foreign key relationship honoring the timepoint of the
+    original query.  This only happens when both tables involved
+    are historic.
+
+    NOTE: related_name has to be different than the one used in
+          TestParticipantToHistoricOrganization as they are
+          sharing the same target table.
+    """
+
+    name = models.CharField(max_length=15, unique=True)
+    organization = HistoricForeignKey(
+        TestOrganizationWithHistory,
+        on_delete=CASCADE,
+        related_name="historic_participants",
+    )
+    history = HistoricalRecords()