+++ simple_history/tests/tests/test_templatetags.py
@@ -3,7 +3,7 @@ from django.test import TestCase
 from simple_history.templatetags.getattributes import getattribute
 
 
-class Foo(object):
+class Foo:
     bar = "bar"