+++ CHANGES.rst
@@ -25,9 +25,14 @@ Full list of changes:
 - Added pre-commit for better commit quality (gh-896)
 - Added ability to break into debugger on unit test failure (gh-890)
 - Russian translations update (gh-897)
+- Fix bug with ``history.diff_against`` with non-editable fields (gh-923)
+- Added HistoricForeignKey (gh-940)
 - Add Python 3.10 to test matrix (gh-899)
 - Added support for Django 4.0 (gh-898)
-- Dropped support for Python 3.6, which reached end-of-life on 2021-12-23 (gh-946).
+- Dropped support for Django 3.1 (gh-952)
+- Dropped support for Python 3.6, which reached end-of-life on 2021-12-23 (gh-946)
+- RecordModels now support a ``no_db_index`` setting, to drop indices in historical models, default stays the same (gh-720)
+- Support change reason formula feature. Change reason formula can be defined by overriding ``get_change_reason_for_object`` method after subclassing ``HistoricalRecords``
 
 3.0.0 (2021-04-16)
 ------------------